backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
stripe==7.8.0
psycopg2-binary==2.9.7
python-decouple==3.8
gunicorn==21.2.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
gspread
